sinkv2(ticdc): move kafka into sink package
