+++ cdc/sinkv2/ddlsink/mq/ddlproducer/ddl_producer.go
@@ -18,7 +18,7 @@ import (
 
 	"github.com/Shopify/sarama"
 	"github.com/pingcap/tiflow/cdc/sink/mq/codec"
-	"github.com/pingcap/tiflow/pkg/kafka"
+	"github.com/pingcap/tiflow/pkg/sink/kafka"
 )
 
 // DDLProducer is the interface for DDL message producer.
