+++ cdc/sinkv2/ddlsink/mq/ddlproducer/kafka_ddl_mock_producer.go
@@ -19,7 +19,7 @@ import (
 	"github.com/Shopify/sarama"
 	mqv1 "github.com/pingcap/tiflow/cdc/sink/mq"
 	"github.com/pingcap/tiflow/cdc/sink/mq/codec"
-	"github.com/pingcap/tiflow/pkg/kafka"
+	"github.com/pingcap/tiflow/pkg/sink/kafka"
 )
 
 var _ DDLProducer = (*MockDDLProducer)(nil)
