+++ cdc/sinkv2/eventsink/mq/dmlproducer/dml_producer.go
@@ -18,7 +18,7 @@ import (
 
 	"github.com/Shopify/sarama"
 	"github.com/pingcap/tiflow/cdc/sink/mq/codec"
-	"github.com/pingcap/tiflow/pkg/kafka"
+	"github.com/pingcap/tiflow/pkg/sink/kafka"
 )
 
 // DMLProducer is the interface for message producer.
