+++ cdc/sinkv2/eventsink/mq/dmlproducer/kafka_dml_mock_producer.go
@@ -20,7 +20,7 @@ import (
 	"github.com/Shopify/sarama"
 	mqv1 "github.com/pingcap/tiflow/cdc/sink/mq"
 	"github.com/pingcap/tiflow/cdc/sink/mq/codec"
-	"github.com/pingcap/tiflow/pkg/kafka"
+	"github.com/pingcap/tiflow/pkg/sink/kafka"
 )
 
 var _ DMLProducer = (*MockDMLProducer)(nil)
