+++ cdc/sinkv2/util/mq/kafka_producer_helper.go
@@ -23,7 +23,7 @@ import (
 	"github.com/pingcap/tiflow/cdc/sink/mq/producer/kafka"
 	"github.com/pingcap/tiflow/pkg/config"
 	cerror "github.com/pingcap/tiflow/pkg/errors"
-	pkafka "github.com/pingcap/tiflow/pkg/kafka"
+	pkafka "github.com/pingcap/tiflow/pkg/sink/kafka"
 )
 
 // GetTopic returns the topic name from the sink URI.
