+++ pkg/sink/kafka/client.go
